Add a digest-path parameter to push commands
Docker images in a registry have an immutable identifier generated by
the registry using a hash of the image manifest. This is very useful
for identifying an exact image.

This digest is returned by the registry when an image is pushed and
added to the local image metadata by the docker engine, which is our
opportunity to save it.

See https://success.docker.com/article/images-tagging-vs-digests