Add checkout and ecr-login flags to build-image
This allows the use of build-image directly for when a user wants
to test that images can be successfully built, but does not want
to push them to the registry. ECR login might still be required
in this case if the base image is hosted in the private ECR registry.