fix: skip-when-tags-exist not working
The issue is that a `contains` query in the CLI command to populate
`docker_tag_exists_in_ecr`, outputs `true` or `false` and not `1` or
`0`, which the test compares against. Changing this to check for `true`
means fixes the issue and correctly increments `number_of_tags_in_ecr`
which will later skip the build.

fixes #200