docs: rm description re: AWS CLI installations
As of v9.x, the commands & jobs does not explicitly install AWS CLI.
Rather, it is assumed that the executor chosen would have the AWS CLI already installed.
Alternatively, users can ensure the AWS CLI is installed as part of the `auth` parameter's steps
when using the build_and_push_image job.